pyproject.toml
[project]
name = "myproject"
version = "0.1.0"
description = "Proyecto Python base"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.3.0,<3.0.0",
  "pydantic[email]>=2.0.0",
  "numpy>=1.24.0,<2.0.0",
  "scikit-learn>=1.3.0,<2.0.0",
  "scipy>=1.11.0,<2.0.0",
  "fastparquet>=2024.11.0,<2025.0.0",
  "pyyaml>=6.0.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.16.0",
  "ruff>=0.11.13",
  "black>=25.1.0",
]

ds = [
  "matplotlib>=3.10.3,<4.0.0",
  "prophet>=1.1.0,<2.0.0",
  "catboost>=1.2.0,<2.0.0",
  "shap>=0.44.0,<1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 119
target-version = ["py311"]

[tool.ruff]
line-length = 119
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

per-file-ignores = { "__init__.py" = ["F401"] }

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
